requirements.txt
Django==1.11.6
Markdown==2.6.9
Pillow==5.0.0
PyYAML==3.12
argparse==1.2.1
certifi==2017.7.27.1
chardet==3.0.4
diff-match-patch==20121119
django-filter==1.0.4
django-import-export==1.0.0
djangorestframework==3.7.0
et-xmlfile==1.0.1
idna==2.6
jdcal==1.3
odfpy==1.3.6
openpyxl==2.5.0
pytz==2017.2
requests==2.18.4
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
wsgiref==0.1.2
xlrd==1.1.0
xlwt==1.3.0
